ajuste de testes
